Added toggle to allow for updates using events

+++ server/src/main/java/edu/unh/cs/cs619/bulletzone/model/events/SpawnEvent.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class SpawnEvent extends GameEvent {
     @JsonProperty
-    private long rawServerValue;
+    private int rawServerValue;
     @JsonProperty
     private int position;
 
     public SpawnEvent() {}
 
     public SpawnEvent(int rawServerValue, int pos) {
-        //super(GameEventType.Spawn);
         this.rawServerValue = rawServerValue;
         this.position = pos;
     }
